requirements.txt
pandas==2.1.4
python_dateutil==2.8.2
streamlit==1.29.0
